+++ src/app/(REST)/get-method/sequential-rendering/page.tsx
@@ -1,8 +1,8 @@
 import { Suspense } from 'react';
 
 import Album from 'components/atomic-design/organisms/Album';
-import { getData } from '@/src/services/data';
-import { getAlbum, getArtist } from '@/src/services/music';
+import { getData } from '@/src/frontend-api/data';
+import { getAlbum, getArtist } from '@/src/frontend-api/music';
 
 // To improve the user experience, you can add a suspense boundary to break up the rendering work and show fetch results in the desired sequence
