+++ src/frontend-api/music.ts
